refactor all layouts into a single layouts.c file
