+++ layouts.h
@@ -1,0 +1,15 @@
+#ifndef LAYOUTS_H
+#define LAYOUTS_H
+
+#include "instantwm.h"
+
+void bstack(Monitor *m);
+void bstackhoriz(Monitor *m);
+void deck(Monitor *m);
+void grid(Monitor *m);
+void monocle(Monitor *m);
+void overviewlayout(Monitor *m);
+void tcl(Monitor * m);
+void tile(Monitor *m);
+
+#endif
